complete API for auth, posts and users

+++ server/routes/post.ts
@@ -4,24 +4,53 @@ import { getAllPosts, getPost, addPost, changePost, deletePost } from '../db/db'
 const post = express.Router();
 
 post.get('/', async (req: any, res: any) => {
-    const result = await getAllPosts();
-    res.json(result);
+    const posts = await getAllPosts();
+    res.json(posts);
 });
 post.get('/:id', async (req: any, res: any) => {
-    const result = await getPost(req.params.id)
-    res.json(result)
+    const post = await getPost(req.params.id)
+    if (post) {
+        res.json(post)
+    } else {
+        res.status(401).json({ error: 'Поста с таким ID нет' })
+    }
 });
 post.post('/', async (req: any, res: any) => {
-    const result = await addPost(req.body.userId, req.body.title, req.body.post)
+    const result = await addPost(req.userId, req.body.title, req.body.post)
     res.json(result);
 });
 post.put('/:id', async (req: any, res: any) => {
-    const result = await changePost(req.params.id, req.body)
-    res.json(result);
+    const post = await getPost(req.params.id);
+    console.log(req.userId);
+    if (post) {
+        if (req.userId == post.userId) {
+            const result = await changePost(req.params.id, req.body)
+            if (result) {
+                const post = await getPost(req.params.id)
+                res.json(post);
+            } else {
+                res.status(401).json({ error: 'Ошибка' })
+            }
+        } else {
+            res.status(401).json({ error: 'Отказано в доступе' })
+        }
+    } else {
+        res.status(401).json({ error: 'Поста с таким ID нет' })
+    }
+
 });
 post.delete('/:id', async (req: any, res: any) => {
-    const result = await deletePost(req.params.id)
-    res.json(result);
+    const post = await getPost(req.params.id);
+    if (post) {
+        if (req.userId == post.userId) {
+            const result = await deletePost(req.params.id)
+            res.json(result);
+        } else {
+            res.status(401).json({ error: 'Отказано в доступе' })
+        }
+    } else {
+        res.status(401).json({ error: 'Поста с таким ID нет' })
+    }
+
 });
 export { post };
-
